removed logedin admin from users list

+++ admin/includes/view_all_users.php
@@ -35,20 +35,23 @@ if ($_SESSION['user_role'] == 'admin') {
         $user_role = escape($row['user_role']);
         $user_image = escape($row['user_image']);
 
-        echo "<tr>";
-        echo "<td> $user_id</td>";
-        echo "<td><img src='../users_images/$user_image' width='50' </td>";
-        echo "<td> $username</td>";
-        echo "<td> $user_firstname</td>";
-        echo "<td> $user_lastname</td>";
-        echo "<td> $user_email</td>";
-        echo "<td> $user_role</td>";
-        echo "<td><a href='users.php?source=edit_user&edit_user=$user_id'>Edit</a></td>";
-        echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
-        echo "<td><a href='users.php?to_admin=$user_id'>Admin</a> <br>";
-        echo "<a href='users.php?to_subs=$user_id'>Subscriber</a></td>";
+        if ($username != $_SESSION['username']) {
 
-        echo "</tr>";
+            echo "<tr>";
+            echo "<td> $user_id</td>";
+            echo "<td><img src='../users_images/$user_image' width='50' </td>";
+            echo "<td> $username</td>";
+            echo "<td> $user_firstname</td>";
+            echo "<td> $user_lastname</td>";
+            echo "<td> $user_email</td>";
+            echo "<td> $user_role</td>";
+            echo "<td><a href='users.php?source=edit_user&edit_user=$user_id'>Edit</a></td>";
+            echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+            echo "<td><a href='users.php?to_admin=$user_id'>Admin</a> <br>";
+            echo "<a href='users.php?to_subs=$user_id'>Subscriber</a></td>";
+
+            echo "</tr>";
+        }
     }
     ?>
     <?php
